Match dll_3_func_5698, dll_3_func_5A48, dll_3_func_6620

+++ include/sys/map.h
@@ -569,7 +569,7 @@ s32 map_get_type(void);
 void map_func_80048034(void);
 void map_func_80048054(s32 mapID, s32, f32 *, f32 *, f32 *, s8 *);
 void map_func_800483BC(f32, f32, f32);
-s8 map_get_layer(void);
+s32 map_get_layer(void);
 
 // something else
